Module Tezos_version_value.Bin_version

val version_string : Tezos_version_parser.t -> string

version_string version returns the version string containing the abbreviated hash, committer date and the current version.

This string is suitable for binaries to show when e.g. a --version flag is given.

val simple_version_string : Tezos_version_parser.t -> string

version_string version returns the version string containing only the current version and the abbreviated hash.

val octez_version_string : string

version_string applied to the Version of Octez.

val octez_simple_version_string : string

simple_version_string applied to the Version of Octez.

val octez_evm_node_version_string : string

version_string applied to the Version of the Octez EVM node.

val octez_evm_node_simple_version_string : string

simple_version_string applied to the Version of Octez EVM node.