Module Tezos_version_value.Current_git_info

val commit_hash : string

val abbreviated_commit_hash : string

val committer_date : string

val octez_version : Tezos_version.Version.t

Current version of the node octez node. it uses either the git tag or a default version

val octez_evm_node_version : Tezos_version.Version.t

Current version of the Octez EVM node. it uses either the git tag or a default version